Originally, the interpreter would exit the execution loop if the instruction was a register jump, which posed a critical performance issue. To address this problem, we now search for the next block at the end of the execution of a register jump. If we find the next block, we link the register jump to it.

As shown in below analysis, this modification improves the overall performance by 1~2%.

* Core i7-11700

| Metric  | Original       |    Proposed    | Speedup |
|---------|----------------|----------------|-------|
| CoreMark| 1931.047 iter/s| 1956.64 iter/s | +1.33%|
| stream  |   79.156 sec   |   77.123 sec   | +2.57%|

Close: #223